user cookie

+++ backend/routes/chat.js
@@ -1,30 +1,31 @@
 import express from 'express'
-import Thread from '../models/Thread.js';
+import { Thread, Messages } from '../models/Thread.js';
 import getGeminiResponse from '../utils/geminai.js';
 
 const router = express.Router();
 
 //test
-router.post("/test", async (req, res) => {
-    try {
-        const thread = new Thread({
-            threadId: "1212abcd",
-            title: "testing11111"
-        })
-        const response = await thread.save()
-        res.send(response);
-    } catch (e) {
-        console.log(e);
-        res.status(500).json({ error: "failed to save data" });
-    }
-})
+//router.post("/test", async (req, res) => {
+//    try {
+//        const thread = new Thread({
+//            threadId: "1212abcd",
+//            title: "testing11111"
+//        })
+//        const response = await thread.save()
+//        res.send(response);
+//    } catch (e) {
+//        console.log(e);
+//        res.status(500).json({ error: "failed to save data" });
+//    }
+//})
 
 //get all thread
 router.get("/thread", async (req, res) => {
     try {
-        const threads = await Thread.find({}).sort({ updatedAt: -1 });
+        const userId = req.userId;
+        const threads = await Thread.find({ cookieId: userId }).sort({ updatedAt: -1 });
         //descending order of updated at
-        res.json(threads);
+        res.json(threads);1
     } catch (e) {
         console.log(e);
         res.status(500).json({ error: "failed to fetch thread" });
@@ -33,14 +34,19 @@ router.get("/thread", async (req, res) => {
 
 router.get("/thread/:threadId", async (req, res) => {
     const { threadId } = req.params;
+    const userId = req.userId;
     try {
-        const thread = await Thread.findOne({ threadId });
+        const thread = await Thread.findOne({ threadId, cookieId: userId });
 
         if (!thread) {
-            res.status(404).json({ error: "Thread not found" });
+            return res.status(404).json({ error: "Thread not found" });
         }
 
-        res.json(thread.messages);
+        //console.log(thread);
+
+        const messages = await Messages.find({ threadId: thread._id }).sort({ createdAt: 1 });
+
+        res.json(messages);
     } catch (e) {
         console.log(e);
         res.status(500).json({ error: "failed to fetch chat" });
@@ -50,12 +56,17 @@ router.get("/thread/:threadId", async (req, res) => {
 
 router.delete("/thread/:threadId", async (req, res) => {
     const { threadId } = req.params;
+    const userId = req.userId;
     try {
-        const thread = await Thread.findOneAndDelete({ threadId });
+        const thread = await Thread.findOneAndDelete({ threadId, cookieId: userId });
 
         if (!thread) {
-            res.status(404).json({ error: "Thread not found" });
+            return res.status(404).json({ error: "Thread not found" });
         }
+
+        console.log(thread);
+
+        await Messages.deleteMany({ threadId: thread._id });
 
         res.status(200).json({ success: "Thread is deleted succesfully" });
     } catch (e) {
@@ -67,33 +78,31 @@ router.delete("/thread/:threadId", async (req, res) => {
 
 router.post("/chat", async (req, res) => {
     const { threadId, message } = req.body;
+    const userId = req.userId;
 
     if (!threadId || !message) {
-        res.status(400).json({ error: "missing requider field" });
+        return res.status(400).json({ error: "missing requider field" });
     }
     try {
-        let thread = await Thread.findOne({ threadId });
-
+        let thread = await Thread.findOne({ threadId, cookieId: userId });
         if (!thread) {
-            //create new thread
-            thread = new Thread({
+            thread = await Thread.create({
                 threadId,
-                title: message,
-                messages: [
-                    {
-                        role: "user",
-                        content: message
-                    }
-                ]
-            })
-        } else {
-            thread.messages.push({ role: "user", content: message });
+                cookieId: userId,
+                title: message
+            });
         }
 
-        const assistantReply = await getGeminiResponse(message);
-        
+        await Messages.create({
+            threadId: thread._id,
+            role: "user",
+            content: message
+        });
 
-        thread.messages.push({ role: "model", content: assistantReply.candidates[0].content.parts[0].text });
+        const assistantReply = await getGeminiResponse(message);
+
+
+        await Messages.create({ threadId: thread._id, role: "model", content: assistantReply.candidates[0].content.parts[0].text });
 
         thread.updatedAt = new Date();
